+++ app/controllers/collections_controller.ts
@@ -1,4 +1,5 @@
 import Collection from "#models/collection"
+import CollectionPolicy from "#policies/collection_policy"
 import { createCollectionValidator, updateCollectionValidator } from "#validators/collection"
 import { cuid } from "@adonisjs/core/helpers"
 import { HttpContext } from "@adonisjs/core/http"
@@ -13,7 +14,10 @@ export default class CollectionsController {
     return collections.serialize()
   }
 
-  async store({ request }: HttpContext) {
+  async store({ request, bouncer, response }: HttpContext) {
+    if (await bouncer.with(CollectionPolicy).denies('create')) {
+      return response.forbidden('Cannot create a collection that is not owned by the user')
+    }
     const {image, ...payload} = await request.validateUsing(createCollectionValidator)
     const collection = new Collection()
     await image.move(app.makePath('storage/uploads/collections/image'), {
@@ -29,9 +33,12 @@ export default class CollectionsController {
     return collection.serialize()
   }
 
-  async update({ request }: HttpContext) {
+  async update({ request, bouncer, response }: HttpContext) {
     const {image, ...payload} = await request.validateUsing(updateCollectionValidator)
     const collection = await Collection.findOrFail(request.params().id)
+    if (await bouncer.with(CollectionPolicy).denies('edit', collection)) {
+      return response.forbidden('Cannot edit a collection that is not owned by the user')
+    }
     await image.move(app.makePath('storage/uploads/collections/image'), {
       name: `${cuid()}.${image.extname}`
     })
@@ -40,8 +47,11 @@ export default class CollectionsController {
     return collection.serialize()
   }
 
-  async destroy({ params }: HttpContext) {
+  async destroy({ params, bouncer, response }: HttpContext) {
     const collection = await Collection.findOrFail(params.id)
+    if (await bouncer.with(CollectionPolicy).denies('delete', collection)) {
+      return response.forbidden('Cannot delete a collection that is not owned by the user')
+    }
     await collection.delete()
   }
 }
